Move icon/clouds/ to clouds/icon/
Convert relative imports to absolute, update icon_physics.py,
parameters.py, and cross-refs in radiation tests.
