+++ jcm/physics/clouds/icon/cloud_microphysics.py
@@ -21,7 +21,7 @@ import jax.numpy as jnp
 from typing import NamedTuple, Tuple, Optional
 import tree_math
 
-from ..constants.physical_constants import (
+from jcm.physics.icon.constants.physical_constants import (
     tmelt, alhf, alhc, alhs, cp, rhow
 )
